+++ package.json
@@ -22,7 +22,6 @@
     "bcryptjs": "^2.4.3",
     "compression": "^1.7.5",
     "cors": "^2.8.5",
-    "@types/cors": "^2.8.17",
     "dotenv": "^16.4.7",
     "express": "^4.21.2",
     "helmet": "^8.0.0",
@@ -37,6 +36,7 @@
     "@types/express": "^5.0.0",
     "@types/helmet": "^4.0.0",
     "@types/jsonwebtoken": "^9.0.7",
+    "@types/cors": "^2.8.17",
     "@types/node": "^22.10.9",
     "nodemon": "^3.1.9",
     "ts-node": "^10.9.2"
